pyproject.toml
[project]
name = "clova-ocr-processor"
version = "0.1.0"
description = "NAVER CLOVA OCR API를 사용한 PDF 텍스트 추출 라이브러리"
readme = "README.md"
license = "MIT"
authors = [
    {name = "chaewonjeong", email = "chaewon@example.com"}
]
keywords = ["ocr", "pdf", "clova", "naver", "text-extraction", "korean-ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.0",
    "pandas>=2.0.0",
    "pymupdf>=1.23.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/chaewonjeong/clova-ocr-processor"

[tool.poetry]
# src 레이아웃 설정
packages = [{include = "clm_ocr", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.0"
pandas = "^2.0.0"
pymupdf = "^1.23.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
